Update_wine_7.11

+++ dlls/ntoskrnl.exe/tests/driver.h
@@ -55,8 +55,8 @@ struct test_data
     int running_under_wine;
     int winetest_report_success;
     int winetest_debug;
-    LONG failures;
-    LONG todo_failures;
+    int failures;
+    int todo_failures;
 };
 
 struct main_test_input
